Background Refactor & Init Fix (#1552)
* Remove relay script

* fix comments

* refactor background & fix messenger issues

* use event instead of timeout

* minor

* fix handlePanel

* fix comments

* fix comments

* minor improvements

* address comments

* Fix minor issue & address comments

---------
