+++ clients/extension/src/content/index.ts
@@ -23,9 +23,8 @@ try {
 }
 const inpageMessenger = initializeMessenger({ connect: 'inpage' })
 
-// There is a 1ms delay for initializing both sides of messengers
-setTimeout(async () => {
+window.addEventListener('vulticonnect:inpage:ready', async () => {
   inpageMessenger.send('setDefaultProvider', {
     vultisigDefaultProvider: await getPrioritizeWallet(),
   })
-}, 1)
+})
